Chore : 데이터 가져오는 부분 list.js로 이동
